+++ src/WeekendRaytracer.jl
@@ -1,8 +1,8 @@
 module WeekendRaytracer
 
-export Ray, at
-export Image, shoot!
-export Camera, get_ray
+# Image and camera
+export Image, shoot!, save
+export Camera
 
 # Hittables
 export Sphere, HittableList
@@ -10,11 +10,14 @@ export Sphere, HittableList
 # Materials
 export Lambertian, Dielectric, Metal
 
+# World generation
+export WorldGeneration
+
 using StaticArrays
 using LinearAlgebra
 using Images
+using Reexport
 using FileIO
-using FLoops
 
 # Inalude abstract types to avoid circular dependencies
 include("AbstractTypes.jl")
@@ -32,7 +35,9 @@ include("Camera.jl")
 include("HitRecord.jl")
 
 # Material
-include("Material/Material.jl")
+include("Material.jl")
+
+# The following are depreciated
 #include("Material/AbstractMaterial.jl")
 #include("Material/Lambertian.jl")
 #include("Material/Dielectric.jl")
@@ -42,6 +47,7 @@ include("Material/Material.jl")
 include("Hittable/Hittable.jl")
 include("Hittable/Sphere.jl")
 include("Hittable/HittableList.jl")
+include("Hittable/WorldGeneration.jl")
 
 # Image
 include("Image.jl")
